requirements.txt
docx==0.2.4
docx2txt==0.8
langchain==0.1.16
langchain_community==0.0.33
langchain_openai==0.1.3
pypdf==4.2.0
python-dotenv==1.0.1
python_docx==1.1.0
streamlit==1.32.2